Fixed overflow in ts/tv calculation on some platforms
Some platforms do not use 64-bit integers for timespec/timeval members. Casting to uint64_t before large multiplication operations. (Hat-tip: Cristian)

+++ pevents.cpp
@@ -90,10 +90,10 @@ namespace neosmart
 				timeval tv;
 				gettimeofday(&tv, NULL);
 				
-				uint64_t nanoseconds = tv.tv_sec * 1000 * 1000 * 1000 + milliseconds * 1000 * 1000 + tv.tv_usec * 1000;
+				uint64_t nanoseconds = ((uint64_t) tv.tv_sec) * 1000 * 1000 * 1000 + milliseconds * 1000 * 1000 + ((uint64_t) tv.tv_usec) * 1000;
 				
 				ts.tv_sec = nanoseconds / 1000 / 1000 / 1000;
-				ts.tv_nsec = (nanoseconds - ts.tv_sec * 1000 * 1000 * 1000);
+				ts.tv_nsec = (nanoseconds - ((uint64_t) ts.tv_sec) * 1000 * 1000 * 1000);
 			}
 			
 			do
@@ -191,10 +191,10 @@ namespace neosmart
 			timeval tv;
 			gettimeofday(&tv, NULL);
 			
-			uint64_t nanoseconds = tv.tv_sec * 1000 * 1000 * 1000 + milliseconds * 1000 * 1000 + tv.tv_usec * 1000;
+			uint64_t nanoseconds = ((uint64_t) tv.tv_sec) * 1000 * 1000 * 1000 + milliseconds * 1000 * 1000 + ((uint64_t) tv.tv_usec) * 1000;
 			
 			ts.tv_sec = nanoseconds / 1000 / 1000 / 1000;
-			ts.tv_nsec = (nanoseconds - ts.tv_sec * 1000 * 1000 * 1000);
+			ts.tv_nsec = (nanoseconds - ((uint64_t) ts.tv_sec) * 1000 * 1000 * 1000);
 		}
 		
 		while(!done)
